requirements.txt
asttokens==2.2.1
backcall==0.2.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
cmake==3.26.3
comm==0.1.3
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dpcpp-cpp-rt==2023.0.0
executing==1.2.0
filelock==3.11.0
fonttools==4.39.3
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-opencl-rt==2023.0.0
intel-openmp==2023.0.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lit==16.0.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mkl==2023.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numexpr==2.8.4
numpy==1.22.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
progressbar2==4.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.6
pycparser==2.21
Pygments==2.15.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.7
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.0
sentry-sdk==1.19.1
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tbb==2021.9.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
ultralytics==8.0.78
urllib3==1.26.14
wcwidth==0.2.6
zipp==3.15.0
